pyproject.toml
[project]
name = "todolist"
version = "0.1.0"
description = ""
authors = [
    {name = "joaovitorpiccolo",email = "joaovitorpiccolo015@gmail.com"}
]
readme = "README.md"
requires-python = "3.12.*"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "todolist", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
run = "fastapi dev src/todolist/main.py"
test = "pytest -s -x --cov=todolist -vv"
uvicorn = "uvicorn todolist.main:app --port 8001"
htmlcov = 'start "C:\\Users\\janj\\AppData\\Local\\Programs\\Opera GX\\opera.exe" C:\\Users\\janj\\PycharmProjects\\todolist\\htmlcov\\index.html'
post_test = "covarage html"
create_table = 'python -c "from src.todolist.database import create_table; create_table()"'
